EA-2193 Move ScaffoldingTestFieldExtractor to test directory
Test-only code does not belong in the production source tree.
Moved from src/dataLayer/clickHouse/fieldExtractors/ to
src/tests/clickhouseOnly/ where it is consumed.
